BISERVER-4509: Reorganizing the parameter processing of PRD/Report-Viewer.
The report viewer no longer processes parameters locally, everything gets passed blindly to the server. The parameter-xml is the only source of configuration the GWT application gets. 

GWT side: We no longer use raw DOM trees to configure everything. Using DOMs is cumbersome and easy to mess up. A real object model exists now that accepts the new xml structure that is generated by the content generator. 

Cleaned out loads of misaligned anonymous inner classes - its never good to use inner classes and then still have to rely on If-then-else streaks if the same can be achieved with smaller classes.

All parameter names are URL-encoded now. Users outside the english language space tend to use their native character set for everything, including URL components.

Instead of hardcoding the error text color, lets use the stylesheet for it.

Server side:
Added proper GWT client support by passing all client properties through the parameter system. Output targets are now used instead of output types. Output targets can be redefined by the user by declaring a parameter with the name of "output-target". Paginating parameter is no longer treated specially, the pagination mode will be activated if the HTML-pageable output target is selected.

Autosubmit, autosubmitui and layout are part of the master-report object now. Reports can be hidden. 

Split the monolithic ContentGenerator into 4 separate classes, each dealing with exactly one request type. Separation of concerns is cool.  The various inner classes that could have been static are now top-level classes. 

If no output target is selected, the default for XActions is HTML-stream, for content generator it is HTML-page. 

Staging-Handler: The configuration for this can now be done via the master-report (in addition to the existing URL or system default settings). If the staging mode is set to memory or temp-file, we now disable the ForceBufferedWriting mode of the HTML export. Be aware that staging in the current implementation causes extra memory consumption for output types that could have been streamed to the user (ie all except bufferd HTML). 

Open points: 
The parameter attributes passed to the client side are not namespace aware yet. The date parameter handling must still be fixed as it is broken for time and timestamp types in Bi-Server 3.5.2 and 3.6.0.