Filter out long running JUnit tests for CI runs
When running an ordinary junit test target, these tests are now
executed without the golden samples. The tests return immediately
when the system property named
 'org.pentaho.reporting.engine.classic.test.ExecuteLongRunningTest'
is set to false.

We need to create one additional CI run-configuration that runs
the engine multi-build with the 'modules' property set to only
'core,testcases' and a ant target of 'longrun-cobertura'.

This will then run the full test suite including the one-hour
tests, without holding up the build process. At the same time
the normal tests are run as part of the build, so that we catch
most errors early in the game.