BISERVER-10541 - Provide a way to resolve common-ui javascript libraries for other projects
Moving require.js and require-cfg.js out of the assembly-provided js and into common-ui. This required an update to PentahoWebContextFilter.java to now include those files from their new home. Also updated the security settings to whitelist these files as they are now behind the authentication barrier. Without this change the files would not be available until logged in.

 By providing require.js in common-ui, it becomes available as part of the js artifact that is published to artifactory. This makes it ivy-resolvable so projects needing these files for unit testing can have them without need of the platform running.