PDI-12140 - IndexOutOfBoundsException when transitioning between different partition schemas
Fix transformation and BaseStep prepare execution logic.
For partitioning we had a case called 'Swim-lanes'. This is a case when we
have 2 steps partitioned with same partitioning schema. For swim-lines we
used to have N_to_N relationship. This is a special mapping when we used
to avoid extra buffers.
The issue was that when we use 2 steps with same patritioning count - but
different partitioning we used to treat this case as a swim lines by
mistake.
This was lead to IOB exception. This case could be workarounded using one
not partitioned step between.
But since we allow to put 2 differenly patritioned steps in line - we
should correctly handle this case. Even it is not optimal transformation
order lead to excessive rowsets.

This commit brings refactoring for steps rowsets relation discovery and
should fix IOOB exception.